"Applied fix from plugins for revision: 1837857" ------------------------------------------------------------------------ r1837857 | jleroux | 2018-08-11 12:37:40 +0200 (sam. 11 août 2018) | 10 lignes
Fixed: Update Apache Tomcat to 9.0.10 because of CVE-2018-8037
(OFBIZ-10517)

If an async request was completed by the application at the same time as the 
container triggered the async timeout, a race condition existed that could 
result in a user seeing a response intended for a different user. 
An additional issue was present in the NIO and NIO2 connectors that did not 
correctly track the closure of the connection when an async request was 
completed by the application and timed out by the container at the same time.
This could also result in a user seeing a response intended for another user. 
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/branches/release17.12@1837862 13f79535-47bb-0310-9956-ffa450edef68