Improved: Cart summary section distorted (OFBIZ-10385)
Navigate to the storefront.
https://demo-trunk.ofbiz.apache.org/ecommerce/control/main 
Add items to the cart.
Check the cart summary section.
Actual:
Cart summary section getting distorted on adding items to the cart

After an issue with r1841672, Garima Jain fixed it by
"Modifiyng the grid classes so that the UI looks good on large screen desktop 
as well as on medium screen desktop width < 1440px or greater than 1024px.

Thanks: Padmavati Rawat for report, Surak for previous commit and Garima for 
the last fix

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/trunk@1845123 13f79535-47bb-0310-9956-ffa450edef68