"Applied fix from plugins for revision: 1847530  " ------------------------------------------------------------------------ r1847530 | jleroux | 2018-11-27 12:27:39 +0100 (mar. 27 nov. 2018) | 12 lignes
Reverted: User should be redirected to view cart page after adding the product 
from compare product screen
(OFBIZ-10667)

Revert r1847478

There was a problem with this fix. It changes the behaviour of additem when you 
add an item from the main or another page.

The behaviour is defined by the "Always View Cart After Adding An Item." 
checkbox in the cart view.

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/ofbiz-plugins/branches/release17.12@1847531 13f79535-47bb-0310-9956-ffa450edef68